Renamed Root to AppWrapper
