+++ src/components/App/App.tsx
@@ -3,14 +3,14 @@ import React from 'react';
 import ThemeProvider from '@material-ui/styles/ThemeProvider';
 import CssBaseline from '@material-ui/core/CssBaseline';
 
-import Root from '../Root';
+import AppWrapper from '../AppWrapper';
 import theme from '../../styles/theme';
 
 const App: React.FC = () => {
   return (
     <ThemeProvider theme={theme}>
       <CssBaseline />
-      <Root />
+      <AppWrapper />
     </ThemeProvider>
   );
 };
