+++ src/components/App/__tests__/App.test.tsx
@@ -7,7 +7,7 @@ jest.mock('@material-ui/core/CssBaseline', () => 'CssBaseline');
 jest.mock('@material-ui/styles/ThemeProvider', () => 'ThemeProvider');
 jest.mock('@material-ui/core/Typography', () => 'Typography');
 jest.mock('../../../styles/theme.ts', () => {});
-jest.mock('../../Root/', () => 'Root');
+jest.mock('../../AppWrapper', () => 'AppWrapper');
 
 describe('Component - App', () => {
   test('renders correctly', () => {
