+++ src/components/AppWrapper/__tests__/AppWrapper.test.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { create } from 'react-test-renderer';
 
-import Root from '../Root';
+import AppWrapper from '../AppWrapper';
 
 jest.mock('@material-ui/core/Container', () => 'Container');
 
-describe('Component - Root', () => {
+describe('Component - AppWrapper', () => {
   test('renders correctly', () => {
-    const component = create(<Root></Root>);
+    const component = create(<AppWrapper></AppWrapper>);
 
     expect(component.toJSON()).toMatchSnapshot();
   });
