+++ src/components/AppWrapper/index.ts
@@ -1,0 +1,3 @@
+import AppWrapper from './AppWrapper';
+
+export default AppWrapper;
