--- src/components/Root/index.ts
@@ -1,3 +1,0 @@
-import Root from './Root';
-
-export default Root;
